### 2019-02-25 #284 @robbriers
cde package submission
Submitting Author: Name (@robbriers)
Repository:   https://github.com/robbriers/cde
Version submitted:   0.3.0
Version accepted: TBD

---

```
Package: cde
Title: Download Water Framework Directive (WFD) data from the Environment Agency Catchment Data Explorer (CDE) website
Version: 0.3.0
Authors@R: person("Rob", "Briers", email = "r.briers@napier.ac.uk",
  role = c("aut", "cre"), comment = c(ORCID = "0000-0003-0341-1203"))
Description: The cde package facilitates searching and download of the
  WFD-related data for all waterbodies within the Environment Agency area.
  The types of data that can be downloaded are: WFD status classification data
  (which can also be plotted), Reasons for Not Achieving Good (RNAG) status,
  objectives set for waterbodies, measures put in place to improve water
  quality and details of associated protected areas. The site accessed is
  https://environment.data.gov.uk/catchment-planning/. The data are made
  available under the Open Government Licence v3.0
  (https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/).
   graphics,
   curl,
   data.table
```

The package facilitates the download, collation and analysis of Water Framework Directive reporting data from the Environment Agency (England) Catchment Data Explorer website.

Any user who wishes to query, download and analyse water quality-related data available from the Catchment Data Explorer website.

No

    - [x] The package is deposited in a long-term repository with the DOI: https://doi.org/10.5281/zenodo.2577179
