### 2019-02-26 a57e07c
fixes for changes in measures/rnag downloads and gp line length issues
