### 2019-02-26 466956f
more line length fixes
