+++ tests/testthat/test_utils.r
@@ -11,8 +11,12 @@ test_that("download returns a dataframe", {
 
 test_that("zip_download returns a dataframe", {
   # set up download URL for RBD level download
-  test_url<-paste0("http://environment.data.gov.uk/catchment-planning/RiverBasinDistrict/6/classification?item=all&status=all&format=csv")
+  
+  test_url<-paste0("http://environment.data.gov.uk/catchment-planning/",
+                   "RiverBasinDistrict/6/classification?item=all",
+                   "&status=all&format=csv")
   testframe<-zip_download(test_url)
   # check that it outputs a dataframe object
   expect_true(is.data.frame(testframe))
 })
+
