### 2019-02-27 #284 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @robbriers !

Here's the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). You've already run good practice, and the below only has a note about test coverage.

```r
── GP cde ───────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 89% of code lines are covered by test
    cases.

    R/get_objectives.r:112:NA
    R/get_objectives.r:113:NA
    R/get_objectives.r:128:NA
    R/get_pa.r:56:NA
    R/get_rnag.r:106:NA
    ... and 28 more lines
```

Seeking reviewers now 🕐

---

Reviewers:

- xx due date: xx
- xx due date: xx
