### 2019-04-04 #284 @boshek
- I think a little more info on the README about usage would be helpful. Right now you are assuming someone clicks through to the pkgdown page
   - The `get_measures` function is not working for me
   - CONTRIBUTING and CODE_OF_CONDUCT files needs in the repo

>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [X] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

  - Almost. I think, however, the `get_measures` function simply returns an empty dataframe.
  - Examples aren't run.

 4
---

I think this is a great package that leverage data access provided by the UK government. It is certainly within rOpenSci' scope for package. The documentation is very good and easy to follow which effectively lower the barrier to entry for a package user. I have some specific comments which are outlined below.

- On the blog down page, you are linked to the "Getting Started" section even though it is "Get Started". This is a potential source of confusion.

#### Vignette
- I think you could revise this line

  > and use of this package implies acceptance of these licence conditions

  to:

  > and use of the data accessed by this package implies acceptance of these licence conditions
- Consider using `remotes` for `install_github()`
- Consider using full argument names in the getting started materials.

- Consider prefixing every function with `cde_*`
- Consider returning columns as all lower case with underscores instead of decimals
- This file: https://github.com/robbriers/cde/blob/master/R/sysdata.rda. Internal data in the package should be included in the data folder.
- No LICENSE file. The open data license would license the data used but not the code in the package
Here is a good description of how to do this: http://r-pkgs.had.co.nz/data.html. In addition, that data should be license within the repo (if it ships with package).
Here is an example of that: https://github.com/ropensci/tidyhydat/tree/master/data-raw/HYDAT_internal_data
- This chunk of code: https://github.com/robbriers/cde/blob/d0e109a2794a2653b15410926a52d2a1f15469bf/R/utils.r#L37-L51
And others like it could probably benefit from a) some helper functions and b) use of `switch`.
- This worries me: https://github.com/robbriers/cde/blob/d0e109a2794a2653b15410926a52d2a1f15469bf/R/utils.r#L94. Could you instead catch the specific error using `tryCatch()`?
- A helper function called something like `cde_data.table` or something like that would be good as you are replicating lots of code here. (https://github.com/robbriers/cde/blob/d0e109a2794a2653b15410926a52d2a1f15469bf/R/utils.r#L94-L114)
- I think you should consider using `data.table`'s progress bar. When downloads are happening, it is ambiguous to the user if the process is working.
- Consider developing print methods (and a class) for cde objects that only print out 10 rows (or just import tibble and use those print methods)
   - I think this could also benefit from `plot` and maybe even `summary` methods.
- Consider letting your examples run so that they are tested with CI
