### 2019-04-05 #284 @robbriers
Thanks for the detailed comments - I'll review and start working through.
