+++ index.Rmd
@@ -21,7 +21,7 @@ knitr::opts_chunk$set(
 
 The types of data that can be downloaded are: WFD status classification data (which can also be plotted), Reasons for Not Achieving Good (RNAG) status, objectives set for waterbodies, measures put in place to improve water quality and details of associated protected areas.
 
-The site accessed is [https://environment.data.gov.uk/catchment-planning/](https://environment.data.gov.uk/catchment-planning/). The data that are downloaded are made available under the Open Government Licence v3.0 (https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/) and use of this package implies acceptance of these licence conditions.
+The site accessed is [https://environment.data.gov.uk/catchment-planning/](https://environment.data.gov.uk/catchment-planning/). The data that are downloaded are made available under the Open Government Licence v3.0 (https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/) and use of the data accessed by this package implies acceptance of these licence conditions.
 
 ## Installation
 
