### 2019-04-09 0e4da8a
Updating docs and starting on utils.r changes. Measures still not being returned for anywhere.
