+++ README.Rmd
@@ -28,8 +28,8 @@ The site accessed is [https://environment.data.gov.uk/catchment-planning/](https
 You can install cde from github with:
 
 ```{r gh-installation, eval = FALSE}
-# install.packages("devtools")
-devtools::install_github("robbriers/cde")
+# install.packages("remotes")
+remotes::install_github("robbriers/cde")
 ```
 
 ## Basic usage
