+++ vignettes/cde.Rmd
@@ -22,7 +22,7 @@ knitr::opts_chunk$set(
 
 The types of data that can be downloaded are: WFD status classification data (which can also be plotted), Reasons for Not Achieving Good (RNAG) status, objectives set for waterbodies, measures put in place to improve water quality and details of associated protected areas.
 
-The site accessed is [https://environment.data.gov.uk/catchment-planning/](https://environment.data.gov.uk/catchment-planning/). The data that are downloaded are made available under the Open Government Licence v3.0 (https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/) and use of this package implies acceptance of these licence conditions.
+The site accessed is [https://environment.data.gov.uk/catchment-planning/](https://environment.data.gov.uk/catchment-planning/). The data that are downloaded are made available under the Open Government Licence v3.0 (https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/) and use of the data accessed by this package implies acceptance of these licence conditions.
 
 ## Installation
 
@@ -30,8 +30,8 @@ You can install cde from github with:
 
 ```{r gh-installation, eval = FALSE}
 # if you have not done so already
-# install.packages("devtools")
-devtools::install_github("robbriers/cde")
+# install.packages("remotes")
+remotes::install_github("robbriers/cde")
 ```
 
 ## Searching for sites
@@ -43,7 +43,7 @@ The `search_sites` function allows you to search for waterbodies, Operational or
 library(cde)
 
 # search for waterbodies containing the name "Lark"
-lark_wb<-search_names("Lark", "name")
+lark_wb<-search_names(string="Lark", column="name")
 ```
 
 The dataframe returned contains details of all the waterbodies containing the string "Lark". The details include waterbody id codes (WBID), type of waterbody, Operational and Management Catchment names and River Basin District.
@@ -55,7 +55,7 @@ head(lark_wb)
 To search for Operational Catchments containing the same string we would use the following code.
 
 ```{r lark search, eval=TRUE}
-lark_oc<-search_names("Lark", "OC")
+lark_oc<-search_names(string="Lark", column="OC")
 ```
 
 ## Retrieving quality status classification data
@@ -85,7 +85,7 @@ head(lark_wb)
 
 # we will get data for the first waterbody here (WBID: GB105033042920, 
 # name: Lark (US Hawstead))
-lark_hawstead<-get_status("GB105033042920", "WBID")
+lark_hawstead<-get_status(ea_name="GB105033042920", column="WBID")
 
 # the dataframe returned contains all of the data for this site in all 
 # years (we did not specify year/year range) (truncate to fit screen)
@@ -98,13 +98,13 @@ An example of a higher level download, specifying a year range and type (in this
 
 ```{r lark river,eval=TRUE}
 # download status data for a given year range and type of waterbody
-lark_OC_rivers<-get_status("Lark", "OC", startyr=2013, endyr=2015, type="River")
+lark_OC_rivers<-get_status(ea_name="Lark", column="OC", startyr=2013, endyr=2015, type="River")
 ```
 To get information about status classification in relation to a specific level in the classification, we can specify `level` as well (see table above for options).
 
 ```{r lark rivers chem, eval=TRUE}
 # download Chemical status for rivers in all years
-lark_OC_rivers_chem<-get_status("Lark", "OC", type="River", level="Chemical")
+lark_OC_rivers_chem<-get_status(ea_name="Lark", column="OC", type="River", level="Chemical")
 ```
 
[…]
