### 2019-04-09 a9f0a57
removing ea_name from subset_data. fixes #16
