### 2019-04-09 c213822
adding progress bar to data.table and ditching curl for download.file to keep progress bars consistent
