### 2019-04-09 e8ad2e9
adding in first switch function for end_url
