### 2019-04-13 2ea423c
converted download functions to several helpers using switch(). Still need to fix WBID name checks.
