+++ R/utils.r
@@ -31,90 +31,32 @@
 
 
 download_cde <- function(ea_name = NULL, column = NULL, data_type=NULL) {
-  # set up url components
-  base_url<-"http://environment.data.gov.uk/catchment-planning/"
+  # do search to make sure that name is present
+  #search_names(ea_name, column)
+  # check that this thows out if error given
   
-  # need to set the end URL differently depending on data_type
-  end_url<-set_end_url(data_type)
-
-  if (column == "RBD") {
-    # rbd level extraction
-    index_num <- ea_wbids$RBD.num[which(ea_wbids[, column] == ea_name)][1]
-    if (is.na(index_num)) {
-      stop("River Basin District name specified not found.")
-    } else {
-      downloadurl <- paste0(base_url, "RiverBasinDistrict/", index_num, 
-                            end_url)
-      cde_data <- zip_download(downloadurl)
-    }
-  } # end of rbd extraction
-  if (column == "MC") {
-    # mc level extraction
-    index_num <- ea_wbids$MC.num[which(ea_wbids[, column] == ea_name)][1]
-    if (is.na(index_num)) {
-      stop("Management Catchment name specified not found.")
-    } else {
-      downloadurl <- paste0(base_url, "ManagementCatchment/", index_num, 
-                            end_url)
-      cde_data <- zip_download(downloadurl)
-    }
-  } # end of mc extraction
-  # oc next
-  if (column == "OC") {
-    # oc level extraction - works
-    index_num <- ea_wbids$OC.num[which(ea_wbids[, column] == ea_name)][1]
-    if (is.na(index_num)) {
-      stop("Operational catchment name specified not found.")
-    } else {
-      cde_data <- data.table::fread(paste0(base_url, "OperationalCatchment/", 
-          index_num, end_url), showProgress = TRUE, header = TRUE, 
-          stringsAsFactors = FALSE, check.names=TRUE, data.table=FALSE)
-    }
-  } # end of oc extraction
-  # finally wbid
-  if (column == "WBID") {
-    # wbid level extraction
-    if (ea_name %in% ea_wbids[, "WBID"]) {
-      if (data_type=="rnag"){
-        # have to add supress warnings as data.table does not like empty 
-        # RNAG data (bad download format on the part of EA)
-        suppressWarnings(cde_data <- data.table::fread(paste0(base_url, 
-          "data/reason-for-failure.csv?waterBody=", ea_name, "&_view=csv"), 
-          showProgress = TRUE, header = TRUE, stringsAsFactors = FALSE, 
-          check.names=TRUE, data.table=FALSE))
-      }
-      if (data_type=="objectives"){
-        cde_data <- data.table::fread(paste0(base_url, "so/WaterBody/", 
-            ea_name, "/objective-outcomes.csv?_view=csv"), 
-            showProgress = TRUE, header = TRUE, stringsAsFactors = FALSE, 
-            check.names=TRUE, data.table=FALSE)
-      }
-      if (data_type=="pa"){
-        cde_data <- data.table::fread(paste0(base_url, "WaterBody/", 
-            ea_name, "/pa/csv"), showProgress = TRUE, header = TRUE, 
-            stringsAsFactors = FALSE, check.names=TRUE, data.table=FALSE)
-      }
-      if (data_type=="class"){
-        cde_data <- data.table::fread(paste0(base_url, 
-            "data/classification.csv?waterBody=", ea_name, "&_view=csv"), 
-            showProgress = TRUE, header = TRUE, stringsAsFactors = FALSE, 
-            check.names=TRUE, data.table=FALSE)
-      }
-    }
-    else {
-      stop("WBID value specified not found.")
-    }
+  # this gives either index number (RBD, MC, OC) or wbid for next bit
[…]
