### 2019-04-15 210e059
tidying up functions - still need to work through check_names
