### 2019-04-15 ba13f03
updating search_names to include WBID column
