+++ R/search_names.r
@@ -1,15 +1,16 @@
 #' Search database of site names
 #' @description Searches the listing of EA monitoring sites to find rows
-#' that contain the string provided. Can search by name (\code{name}),
-#' Management Catchment (\code{MC}), Operational Catchment (\code{OC})
-#' or River Basin District (\code{RBD}). The search is done on a local copy
-#' of the waterbody listing rather than connecting to the EA site.
+#' that contain the string provided. Can search by WBID (\code{WBID}), name 
+#' (\code{name}), Management Catchment (\code{MC}), Operational Catchment 
+#' (\code{OC}) or River Basin District (\code{RBD}). The search is done 
+#' on a local copy of the waterbody listing rather than connecting to the 
+#' EA site.
 #
 #' @param string The search string to be matched. Will match whole or partial
 #' strings in the column values.
 #'
 #' @param column The column to be searched. Possible options are
-#' \code{name}, \code{OC} (Operational Catchment), \code{MC}
+#' \code{WBID}, \code{name}, \code{OC} (Operational Catchment), \code{MC}
 #' (Management Catchment) and \code{RBD} (River Basin District)
 #'
 #' @return A data frame containing the details of all the sites that match
@@ -25,10 +26,10 @@
 #' \dontrun{search_names("Kennet", "MC")}
 #' 
 search_names <- function(string = NULL, column = NULL) {
-  search_choices <- c("name", "MC", "OC", "RBD")
+  search_choices <- c("name","WBID", "MC", "OC", "RBD")
   # check if there is a value passed for both arguments
   if (is.null(column) | is.null(string)) {
-    stop("Both a search string and column (name, MC, OC, or RBD) 
+    stop("Both a search string and column (name, WBID, MC, OC, or RBD) 
          should be specified", "\n")
   }
   # if the column is found in ea_wbids
@@ -38,7 +39,7 @@ search_names <- function(string = NULL, column = NULL) {
     # remove indexing columns
     matching_rows <- matching_rows[, c(-5, -7, -9)]
   } else {
-    stop("Column specified should be: name, MC, OC, or RBD", "\n")
+    stop("Column specified should be: name, WBID, MC, OC, or RBD", "\n")
   }
   #  }
   if (nrow(matching_rows) == 0) {
