+++ DESCRIPTION
@@ -20,7 +20,7 @@ Encoding: UTF-8
 LazyData: true
 Imports:
    graphics,
-   curl,
+   utils,
    data.table
 Suggests:
     testthat,
