### 2019-04-15 #284 @robbriers
@boshek Just a query re license and data. The code is licensed as GPL 3 in the DESCRIPTION file, so I was under the impression that this did not need an explicit LICENSE file to be included?
Re the data, I was intending to keep the data internal to the package, so as per the Hadley Packages site: 'If you want to store parsed data, but not make it available to the user, put it in R/sysdata.rda'. However now I am wondering whether as the internal data were derived from the Open Data source, the licence for this needs to be explicitly stated somewhere?
