### 2019-04-15 #284 @boshek
> The code is licensed as GPL 3 in the DESCRIPTION file, so I was under the impression that this did not need an explicit LICENSE file to be included

I don't think it is absolutely required but I it is a) a [generally](https://github.com/RcppCore/Rcpp) [accepted](https://github.com/tidyverse/dplyr) practice to include a LICENSE file in the repo and b) provides a path for a non R user (not to mention GitHub) to look at the license. That is, someone wanting to find out how this software is licensed only needs to come to repo and doesn't need to know that one needs to look into the DESCRIPTION file.

> Internal data were derived from the Open Data source, the licence for this needs to be explicitly stated somewhere?

It definitely does. My understanding is that regardless of how the data in packaged if you include it you need to identify how it is licensed. I would also encourage you to expose this data. As a prospective user I found it confusing that I couldn't find out more on this data. That's why I suggested not only exposing the data but also providing a data-raw folder. An example I've done is as follows. Here is how I create and provide licensing info for an internal data file

https://github.com/ropensci/tidyhydat/tree/master/data-raw/HYDAT_internal_data

Then here I document the data for context for folks:

https://github.com/ropensci/tidyhydat/blob/e63704360b4cdac9283752659738553e919ff170/R/data.R#L14-L35

Does that help?
