+++ data-raw/cde_internal_data.Rmd
@@ -1,0 +1,15 @@
+---
+output: github_document
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+## cde internal data
+
+For the purpose of constructing API calls, `cde` makes calls to an internal datadrame (`ea_wbids`) which consists of a table of the details (name and index number for each site/catchment) of all waterbodies that can be downloaded from [https://environment.data.gov.uk/catchment-planning/](https://environment.data.gov.uk/catchment-planning/). The file here (`ea_wbids.csv`) contains a copy of the internal data content. The data are provided under the terms of the Open Government Licence v3.0 (https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/).
