+++ DESCRIPTION
@@ -15,7 +15,7 @@ Description: The cde package facilitates searching and download of the
   (https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/).
 Depends:
     R (>= 3.0.0)
-License: GPL-3
+License: GPL-3 | file LICENCE
 Encoding: UTF-8
 LazyData: true
 Imports:
