+++ NAMESPACE
@@ -8,3 +8,4 @@ export(get_status)
 export(plot_status)
 export(search_names)
 importFrom(graphics,barplot)
+importFrom(utils,download.file)
