### 2019-04-15 #284 @robbriers
Thanks, I have added an explicit licence file.

With the data I would prefer to keep the internal df in the sysdata.rda file, but have added a copy of the data with appropriate licence details to a data-raw folder.
