### 2019-04-17 #284 @robbriers
@boshek if I document and expose the data in /data I seem to run into an issue with ` no visible binding for global variable 'ea_wbids'`. I am sure that I have come across this issue before but can't recall/find how to fix it - any advice?
