### 2019-04-17 bb67b6a
moving ea_wbids to /data but issue with global binding....
