### 2019-04-17 #284 @boshek
What if instead of `ea_wbids.r` the file describing the data was `data.r`? Total shot in the dark.
