+++ man/ea_wbids.Rd
@@ -1,29 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ea_wbids.r
-\docType{data}
-\name{ea_wbids}
-\alias{ea_wbids}
-\title{Details of name and index of all sites/catchments.}
-\format{A data frame with 5237 rows and 9 variables:
-\describe{
-  \item{WBID}{identifier for individual waterbodies}
-  \item{name}{detailed name of the site/catchment}
-  \item{type}{type of waterbody (River, Lake etc.)}
-  \item{OC}{Operational catchment name}
-  \item{OC_num}{Index number of the Operational Catchment}
-  \item{MC}{Management catchment name}
-  \item{MC_num}{Index number of the Management Catchment}
-  \item{RBD}{River Basin District name}
-  \item{RBD_num}{Index number of the River Basin District}
-  ...
-}}
-\source{
-\url{https://environment.data.gov.uk/catchment-planning/}
-}
-\usage{
-ea_wbids
-}
-\description{
-Dataframe used by `cde` to construct API calls
-}
-\keyword{datasets}
