### 2019-04-18 ec8a926
setting up class and print method
