### 2019-04-18 dc6634f
adding class to return values - some issues still with print
