### 2019-04-18 7f79845
updating print to handle empty dfs. Fixes #17
