+++ R/cde-class.r
@@ -8,6 +8,7 @@ as.cde <- function(x) {
 # custom print method
 print.cde_df <- function(x){
   # find number of columns that will fit on current width
+  ### checking no data situation
   cols<-min(which(cumsum(nchar(names(x))+2) > getOption("width")))-1
   
   # subset df for just these columns
@@ -15,16 +16,17 @@ print.cde_df <- function(x){
   
   # get column name lengths for use in truncation
   col_name_lengths<-nchar(names(data_to_print))
-
+  
   # if there are more than 10 rows, just take first 10
   if (nrow(data_to_print)>10){
     data_to_print <- data_to_print[1:10,]
   }
   # truncate strings within rows to fit as well
-  data_to_print <- as.data.frame(t(apply(data_to_print, 1, trunc_char, cols, col_name_lengths)))
-  
+  if(!nrow(x)==0){
+    data_to_print <- as.data.frame(t(apply(data_to_print, 1, trunc_char, cols, col_name_lengths)))
+    print(data_to_print, row.names=FALSE)
+  }else{cat("No data returned - printing not possible")}
   # output data that fits
-  print(data_to_print, row.names=FALSE)
   
   # if more than 10 rows, indicate missing data
   if(nrow(x)>10){
@@ -32,7 +34,7 @@ print.cde_df <- function(x){
     cat(paste0("With an additional ", nrow(x)-10, " rows and ", ncol(x)-ncol(data_to_print), " columns of data."),"\n")
     cat("Row values may be truncated to fit console.")
   }
-  # end of function
+    # end of function
 }
 
 # string truncation function
