### 2019-04-18 e790034
updating print method to account for situation where columns all fit on screen. Still need to fix reporting of extra rows/cols. Also removed cde_df conversion for search_names, need to update column names for consistency.
