### 2019-04-18 f2e77dd
updating output to account for different dimensions and documenting params
