### 2019-04-18 #284 @boshek
@robbriers This is a workaround to `global variables` issue:

```
.onLoad <- function(libname = find.package("cde"), pkgname = "cde") {
  # CRAN Note avoidance
  if (getRversion() >= "2.15.1")
    utils::globalVariables(
      c("ea_wbids")
    )
  invisible()
}
```
Just add that bit of code to your `zzz.R` file.
