### 2019-04-18 cee4959
fixed global variables issue with .onLoad function, so shifting ea_wbids to /data
