### 2019-04-19 #284 @limnoliver
- I think the `get_measures` function is working (did for me below), you've just chosen two examples that don't have any data. Maybe choose one that does? (see my example below)

- see @boshek review

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

- see @boshek review

Estimated hours spent reviewing: 6 (though this was my first time reviewing for ropensci, so a lot of this was getting up to speed on the review process. Thanks for all the great documentation!)

---

Congrats on providing a useful package that allows for easier and reproducible ways to access and use important water data. Most of my comments are for making it easier for someone who is new to the data get up-to-speed more quickly. I don't think this will take too much effort, and could be as simple as linking to some documentation that likely exists on the WFD website. A few comments:

>Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient? Does it use the principle of multiple points of entry i.e. takes into account the fact that any piece of documentation may be the first encounter the user has with the package and/or the tool/data it wraps?

I think the documentation could be improved in regards to describing the data source the package is accessing. In the README, it would be useful to have an extra sentence or two that describes the scope of the data - e.g., data collected by X for the purposes of Y within the U.K. etc. I had never heard of the WFD or the Environment Agency, so I had to take one more step to understand where the data are coming from.

To test the functions, I followed the vignettes and created my own examples, and I show those below with comments on individual functions.

`lakes <- cde::search_names('Lake', 'name')`

- It might be useful to note in the `search_names` documentation that the search is case-sensitive. Also, in the `Value` section of the help file, it would be helpful to the user to see definitions of the columns that are returned.
```
bolam_dat <- get_status(lakes$WBID[lakes$name %in% 'Bolam Lake'], 'WBID')
names(bolam_dat)

##  [1] "River.Basin.District"           "Management.Catchment"
##  [3] "Operational.Catchment"          "Waterbody.ID"
##  [5] "water.body"                     "water.body.type"
##  [7] "easting"                        "northing"
##  [9] "ngr"                            "Hydromorphological.designation"
## [11] "Year"                           "Status"
## [13] "Cycle"                          "Classification.Level"
## [15] "Classification.Item"            "classification.ID"
## [17] "Certainty"                      "Confidence"
## [19] "Type"                           "Deterioration.type"
## [21] "Linked.Reasons"                 "Investigation.outcome"
```

- I agree with the other reviewer that it would be nice to have column names standardized and cleaned up to exclude periods.

- Here as well, it would be useful to have more definitions in the `Values` section of the help file for `get_status`. If you don't feel like it's within the scope of the package to define the columns that are returned, a link to metadata would be useful.

`bolam_oc <- get_status(lakes$OC[lakes$name %in% 'Bolam Lake'], 'OC', level = 'Chemical')`

- I had a hard time understanding "level", and in the `get_status` help file it says to go to the vignette for more info. I would include a direct link to the vignette section of relevance, or spell out the options in the help file.

`
plot_status(lakes$OC[lakes$name %in% 'Bolam Lake'], 'OC')`
- It seems like a common workflow would be to get the status data, and then to plot. To do that easily using the package, the user would use `get_status` and `plot_status`. I was expecting `plot_status` to use the output of `get_status`, but it instead makes a call to `get_status` within the function and and downloads the data again, so if you want to both see the data in a data frame and plot it, you have to download it twice.

`bolam_oc_rnag <- get_rnag(lakes$OC[lakes$name %in% 'Bolam Lake'], 'OC')`

- Again, it would be helpful to have more in-depth description of the returned table, or at least a link to a description for how these classifications are made or descriptions of the columns. (same can be said for `get_objectives`, `get_measures`, and `get_pa`)

```
bolam_oc_meas <- get_measures(lakes$OC[lakes$name %in% 'Bolam Lake'], 'OC')
## No measures data specified - empty dataframe returned
```
- Good warning in the vignette that this often returns no results!
