### 2019-04-22 11ac6d7
convert all columns names to consistent lower case and _ separator. Closes #19.
