### 2019-04-22 #284 @robbriers
@limnoliver thanks for the review - I will go through the comments and address!
