### 2019-04-22 5cfe822
issse with RNAG data - change in format on EA site so need to check level arg
