### 2019-04-23 13eb61a
setting up print method - needs development still.
