+++ man/print.cde_df.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/cde-class.r
+\name{print.cde_df}
+\alias{print.cde_df}
+\title{Print method definition for cde_df}
+\usage{
+\method{print}{cde_df}(x, ...)
+}
+\arguments{
+\item{x}{An object of class `cde_df`}
+
+\item{...}{Other arguments passed on to individual methods}
+}
+\description{
+Custom `print` method for objects of class '`cde_df`.
+Formats output to fit current width of console.
+}
