+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(print,cde_df)
 export(get_measures)
 export(get_objectives)
 export(get_pa)
