### 2019-04-29 29a73ea
working on plot methods - bit commit but still a WIP
