+++ R/get_pa.r
@@ -49,6 +49,6 @@ get_pa <- function(ea_name = NULL, column = NULL) {
     message("No protected areas present - empty dataframe returned")
   }
   # add comment for plot method identification
-  comment(pa_data) <- "pa"
+  comment(pa_data) <- paste0("pa;", column)
   return(as.cde(pa_data))
 } # end of function
