+++ R/plot_status.r
@@ -1,115 +1,115 @@
-#' Plot Status Summary
-#' @description Produces a (stacked) percentage barplot of waterbody
-#' status information for a given set of data (MC, OC or RBD)
-
-#' @inheritParams get_status
-#'
-#' @param scheme Which colour scheme to use with plots; defaults to a viridis
-#' based scheme (\code{"vir"} but can also choose to use the colours specified
-#' in the WFD document by specifying as \code{"wfd"}.
-#'
-#' @importFrom graphics barplot
-#'
-#' @return A (stacked) barplot of the percentage of waterbodies within the
-#' specified area of different status values represented as different colours
-#' depending on the scheme specified.
-#'
-#' @export plot_status
-#'
-#' @examples
-#' # plot the Overall Water Body status for Rivers in the Avon Warwickshire
-#' # Management Catchment in 2011 using viridis-based colours
-#' plot_status("Avon Warwickshire", "MC", startyr = 2011, type = "River")
+#' #' Plot Status Summary
+#' #' @description Produces a (stacked) percentage barplot of waterbody
+#' #' status information for a given set of data (MC, OC or RBD)
 #' 
-#' # Chemical status classification for same MC, but all years and types
-#' plot_status("Avon Warwickshire", "MC", level = "Chemical")
+#' #' @inheritParams get_status
+#' #'
+#' #' @param scheme Which colour scheme to use with plots; defaults to a viridis
+#' #' based scheme (\code{"vir"} but can also choose to use the colours specified
+#' #' in the WFD document by specifying as \code{"wfd"}.
+#' #'
+#' #' @importFrom graphics barplot
+#' #'
+#' #' @return A (stacked) barplot of the percentage of waterbodies within the
+#' #' specified area of different status values represented as different colours
+#' #' depending on the scheme specified.
+#' #'
+#' #' @export plot_status
+#' #'
+#' #' @examples
+#' #' # plot the Overall Water Body status for Rivers in the Avon Warwickshire
+#' #' # Management Catchment in 2011 using viridis-based colours
+#' #' plot_status("Avon Warwickshire", "MC", startyr = 2011, type = "River")
+#' #' 
+#' #' # Chemical status classification for same MC, but all years and types
+#' #' plot_status("Avon Warwickshire", "MC", level = "Chemical")
+#' #' 
+#' #' # plot the Overall Water Body status of all waterbodies
+#' #' # in the Humber RBD between 2012 and 2014 using WFD colour scheme
+#' #' plot_status("Humber", "RBD", startyr = 2012, endyr = 2014, scheme = "wfd")
+#' #' 
+#' plot_status <- function(ea_name = NULL, column = NULL, 
+#'     level = "Overall Water Body", startyr = NULL, endyr = NULL, 
+#'     type = NULL, scheme = "vir") {
+#'   # start by running general checks on input data
+#'   check_args(ea_name, column, startyr, endyr, type)
+#'   # do initial check of column choice
+#'   plot_choices <- c("MC", "OC", "RBD")
+#'   if (!column %in% plot_choices) {
+#'     stop("Column specified is not one of the possible choices 
+#'          (\"OC\", \"MC\" or \"RBD\").")
+#'   }
+#'   # check that scheme is specified correctly
+#'   scheme_choices<-c("vir", "wfd")
+#'   if (!scheme %in% scheme_choices){
+#'     stop("scheme should be either \"vir\" or \"wfd\".")
+#'   }
+#'   # get required data
[…]
