+++ NAMESPACE
@@ -6,6 +6,7 @@ export(get_objectives)
 export(get_pa)
 export(get_rnag)
 export(get_status)
+export(plot.cde_df)
 export(plot_status)
 export(search_names)
 importFrom(graphics,barplot)
