### 2019-04-29 15767cb
removing plots from vignettes as causes issue with CHECK until plot methods finished - some other errors in docs and method export - bit of a mess
