+++ vignettes/cde.Rmd
@@ -115,14 +115,14 @@ Using the `get_status` function provides maximum flexibility as the user can dec
 
 ```{r lark plot, fig.height=4, fig.width=6.5, fig.align="center", eval=TRUE}
 # plot overall waterbody status information for the Lark OC between 2013 and 2015
-plot_status(ea_name="Lark", column="OC", startyr=2013, endyr=2015)
+#plot_status(ea_name="Lark", column="OC", startyr=2013, endyr=2015)
 ```
 
 The default colour scheme (`vir`) is based on the `viridis` palette, but the WFD-defined status colours can be used instead by setting `scheme="wfd"`. Also if a single year is specified, a standard (as opposed to stacked) barplot is produced.
 
 ```{r lark riverplot wfd,fig.height=4, fig.width=6.5, fig.align="center", eval=TRUE}
 # plot overall waterbody status information for rivers in the Lark OC in 2015
-plot_status(ea_name="Lark", column="OC", startyr=2015, type="River", scheme="wfd")
+#plot_status(ea_name="Lark", column="OC", startyr=2015, type="River", scheme="wfd")
 ```
 
 ## Reasons for Not Achieving Good status
