### 2019-04-29 2ca17fd
plots about sorted, but need docs still. fixing issue in print method too
