+++ NAMESPACE
@@ -6,7 +6,6 @@ export(get_objectives)
 export(get_pa)
 export(get_rnag)
 export(get_status)
-export(plot.cde_df)
 export(plot_status)
 export(search_names)
 importFrom(graphics,barplot)
