+++ R/utils.r
@@ -268,7 +268,6 @@ check_args <- function(ea_name = NULL, column = NULL, startyr = NULL,
 #'
 subset_data <- function(full_data, column = NULL, 
     level = "Overall Water Body", startyr = NULL, endyr = NULL, type = NULL) {
-
   # if only start year is set, is it beyond the data range?
   if (!is.null(startyr) & is.null(endyr)){
     if (startyr>max(full_data$year)){
