+++ R/get_rnag.r
@@ -90,6 +90,7 @@ get_rnag <- function(ea_name = NULL, column = NULL, startyr = NULL,
       }
     }
     # subset data
+    print(max(rnag_data$year))
     rnag_data<-subset_data(rnag_data, column, NULL, startyr, endyr, type)
     if (nrow(rnag_data)==0){
       message("No RNAG data - empty dataframe returned")
