+++ NAMESPACE
@@ -1,12 +1,13 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(plot,cde_df)
 S3method(print,cde_df)
 export(get_measures)
 export(get_objectives)
 export(get_pa)
 export(get_rnag)
 export(get_status)
-export(plot_status)
 export(search_names)
 importFrom(graphics,barplot)
+importFrom(graphics,par)
 importFrom(utils,download.file)
