### 2019-04-30 85ce5fd
condensing methods to single doc. need to work on docs still
