+++ R/get_rnag.r
@@ -93,7 +93,6 @@ get_rnag <- function(ea_name = NULL, column = NULL, startyr = NULL,
       }
     }
     # subset data
-    print(max(rnag_data$year))
     rnag_data<-subset_data(rnag_data, column, NULL, startyr, endyr, type)
     if (nrow(rnag_data)==0){
       message("No RNAG data - empty dataframe returned")

[…]
