+++ R/cde-class_and_methods.r
@@ -1,9 +1,10 @@
 #' Class definition for cde_df
 #' @description Output from a call to the EA website is returned as an
-#' object of class '`cde_df`, which is basically a dataframe with a 
-#' modified `print` method to format the output ro fit the console width
+#' object of class \code{cde_df}, which is basically a dataframe with 
+#' modified \code{print} method to format the output to fit the console 
+#' width and a \code{plot} method to produce default plots of the data.
 #
-#' @param x A dataframe to be converted to class `cde_df`
+#' @param x A dataframe to be converted to class \code{cde_df}
 #'
 #'@noRd
 
@@ -13,18 +14,19 @@ as.cde <- function(x) {
 }
 
 #' Print method definition for cde_df
-#' @description Custom `print` method for objects of class '`cde_df`.
+#' @description Custom \code{print} method for objects of class \code{cde_df}.
 #' Formats output to fit current width of console.
 #' 
-#' @param x An object of class `cde_df`
+#' @param x An object of class \code{cde_df}.
 #' 
-#' @param ... Other arguments passed on to individual methods
+#' @param ... Other arguments passed on to individual methods. None 
+#' implemented at present.
 #'
 #' @method print cde_df
 #' @export
 
 print.cde_df <- function(x, ...){
-  # find number of columns that will fit on current width
+  # find number of columns that will fit on current console width
   # if the maximum length of all column names is greater than the width
   # subset the columns
   if(max(cumsum(nchar(names(x))+2)>getOption("width"))){
@@ -44,16 +46,19 @@ print.cde_df <- function(x, ...){
   
   # truncate strings within rows to fit as well
   if(!nrow(x)==0){
-    data_to_print <- as.data.frame(t(apply(data_to_print, 1, trunc_char, col_name_lengths)))
+    data_to_print <- as.data.frame(t(apply(data_to_print, 1, trunc_char, 
+      col_name_lengths)))
     print(data_to_print, row.names=FALSE)
   }else{cat("No data returned - printing not possible")}
   
   # if more than 10 rows, indicate missing data
   if(nrow(x)>10  & ncol(x)>ncol(data_to_print)){
-    cat(paste0("With an additional ", nrow(x)-10, " rows and ", ncol(x)-ncol(data_to_print), " columns of data."),"\n")
+    cat(paste0("With an additional ", nrow(x)-10, " rows and ", 
+      ncol(x)-ncol(data_to_print), " columns of data."),"\n")
   }
   if (nrow(x)<11  & ncol(x)>ncol(data_to_print)){
-    cat(paste0("With an additional ", ncol(x)-ncol(data_to_print), " columns of data."),"\n")
+    cat(paste0("With an additional ", ncol(x)-ncol(data_to_print), 
+      " columns of data."),"\n")
   }
   if(nrow(x)>10 & ncol(x)<=ncol(data_to_print)){
     cat(paste0("With an additional ", nrow(x)-10, " rows of data."),"\n")
@@ -62,14 +67,14 @@ print.cde_df <- function(x, ...){
   # end of function
 }
 
-#' Truncate strings within `cde_df` objects to fit console
-#' @description Truncates the length of strings within rows of `cde_df`
+#' Truncate strings within \code{cde_df} objects to fit console
+#' @description Truncates the length of strings within rows of \code{cde_df}
 #' objects to the same length as the column name, ensuring that they
 #' fit current width of console.
 #' 
-#' @param x An object of class `cde_df`
+#' @param x An object of class \code{cde_df}.
 #' 
-#' @param col_name_lengths Vector containing lengths of column names
[…]
