+++ man/print.cde_df.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cde-class.r
+% Please edit documentation in R/cde-class_and_methods.r
 \name{print.cde_df}
 \alias{print.cde_df}
 \title{Print method definition for cde_df}
@@ -7,11 +7,12 @@
 \method{print}{cde_df}(x, ...)
 }
 \arguments{
-\item{x}{An object of class `cde_df`}
+\item{x}{An object of class \code{cde_df}.}
 
-\item{...}{Other arguments passed on to individual methods}
+\item{...}{Other arguments passed on to individual methods. None 
+implemented at present.}
 }
 \description{
-Custom `print` method for objects of class '`cde_df`.
+Custom \code{print} method for objects of class \code{cde_df}.
 Formats output to fit current width of console.
 }
