+++ man/plot.cde_df.Rd
@@ -1,30 +1,38 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cde-class.r
+% Please edit documentation in R/cde-class_and_methods.r
 \name{plot.cde_df}
 \alias{plot.cde_df}
-\title{Plot method for cde_df output}
+\title{Plot method for \code{cde_df} output}
 \usage{
 \method{plot}{cde_df}(x, scheme = "vir", ...)
 }
 \arguments{
-\item{x}{An object of class `cde_df` to be plotted}
+\item{x}{An object of class \code{cde_df} to be plotted.}
 
-\item{scheme}{Which colour scheme to use with plots; defaults to a viridis
-based scheme (\code{"vir"}) but for \code{rnag} and \code{objectives} data 
-the colours specified in the WFD document by specifying as \code{"wfd"}.}
-
-\item{...}{Other arguments passed on to individual methods}
+\item{...}{Other arguments passed on to individual methods. The only other
+argument implemented at present is \code{scheme}. For \code{status} and 
+\code{objectives} this defines Which colour scheme to use with plots.It 
+defaults to a viridis-based scheme (\code{"vir"}). Alternatively, the 
+colours specified in the WFD document by specifying as \code{"wfd"}.}
 }
 \description{
-Default plots of the output of calls to `get...` functions.
-For `status` and `objectives` produces a (stacked) percentage barplot of 
-waterbody observed or predicted (objective) status information for a 
-given set of data. 
+Default plots of the output of calls to \code{get...}
+ functions. Details of the plots for different data are given below.
+ 
+ \tabular{ccc}{
+ Function \tab Data plotted \tab Plot type \cr
+ \code{get_status} \tab status class (by year if included) \tab percentage barplot \cr
+ 
+ }
+ 
+ For \code{status} and \code{objectives} produces a (stacked) 
+ percentage barplot of waterbody observed or predicted (objective) 
+ status information for a given set of data. 
 
-For `rnag`, `measures` or `pa` produces a frequency histogram of the 
-columns `pressure_tier_3` for `rnag`, `measures_tier_1` for `measures`
+For \code{rnag}, \code{measures} or \code{pa} produces a frequency 
+histogram of the columns `pressure_tier_3` for `rnag`, `measures_tier_1` for `measures`
  and `protected_area_type` for `pa`. Details of the classifications used
- are given XXXXXX
-
+ are given XXXXXX.
+ 
 Plotting is only possible for MC, OC or RBD downloads.
 }
