### 2019-04-30 b6ac38a
updating docs still. need to sort out EA urls
