### 2019-04-30 a1ba760
fixing issues with plots - still need to finish docs
