+++ NAMESPACE
@@ -11,3 +11,4 @@ export(search_names)
 importFrom(graphics,barplot)
 importFrom(graphics,par)
 importFrom(utils,download.file)
+importFrom(viridisLite,viridis)

[…]
