### 2019-05-03 37869f9
more doc updates
