+++ R/get_measures.r
@@ -8,11 +8,13 @@
 #' to a target change in status are included. Therefore in many cases 
 #' the function will return an empty dataframe with a message.
 #
-#' @param ea_name A string representing the description (name) of the
-#' features to be extracted. For example to extract data for the whole of
-#' the Humber RBD, this would be "Humber"; also see examples. Must be an
-#' exact match to the values used in the EA database.
-#' Use the \code{\link{search_names}} function to search for specific values.
+#' @param ea_name A string representing the description (\code{name} for 
+#' \code{OC}, \code{MC} or \code{RBD} level downloads or \code{WBID} for 
+#' individual waterbodies) of the features to be extracted. For example 
+#' to extract data for the whole of the Humber RBD, this would be "Humber"; 
+#' also see examples. Must be anexact match to the values used in the EA 
+#' database. Use the \code{\link{search_names}} function to search for 
+#' specific values.
 #'
 #' @param column The column to be searched. Possible options are
 #' \code{WBID} (waterbody id), \code{OC} (Operational Catchment), \code{MC}
