+++ R/get_pa.r
@@ -6,12 +6,13 @@
 #' (\code{WBID}), Management Catchment (\code{MC}), Operational
 #' Catchment (\code{OC}) or River Basin District (\code{RBD}).
 #
-#' @param ea_name A string representing the description (name for \code{OC}, 
-#' \code{MC} or \code{RBD} level downloads or \code{WBID} for individual 
-#' waterbodies) of the features to be extracted. For example to extract 
-#' data for the whole of the Humber RBD, this would be "Humber"; also 
-#' see examples. Must be an exact match to the values used in the EA database.
-#' Use the \code{\link{search_names}} function to search for specific values.
+#' @param ea_name A string representing the description (\code{name} for 
+#' \code{OC}, \code{MC} or \code{RBD} level downloads or \code{WBID} for 
+#' individual waterbodies) of the features to be extracted. For example 
+#' to extract data for the whole of the Humber RBD, this would be "Humber"; 
+#' also see examples. Must be an exact match to the values used in the 
+#' EA database. Use the \code{\link{search_names}} function to search 
+#' for specific values.
 #'
 #' @param column The column to be searched. Possible options are
 #' \code{WBID} (waterbody id), \code{OC} (Operational Catchment), \code{MC}
