### 2019-05-06 f1a6617
more doc updates
