### 2019-05-06 9431fb3
docs again - allowing examples to run
