+++ R/get_rnag.r
@@ -49,15 +49,15 @@
 #'
 #' @examples
 #' # get all RNAG issues identified for waterbody GB112071065700
-#' \dontrun{get_rnag("GB112071065700", "WBID")}
+#' get_rnag("GB112071065700", "WBID")
 #' 
 #' # get the RNAG issues for Lakes in the Humber RBD, between
 #' # 2013 and 2014
-#' \dontrun{get_rnag("Humber", "RBD", startyr = 2013, endyr = 2014, type = "Lake")}
+#' get_rnag("Humber", "RBD", startyr = 2013, endyr = 2014, type = "Lake")
 #' 
 #' # get the RNAG issues for Rivers in the Avon Warwickshire
 #' # Management Catchment in 2015
-#' \dontrun{get_rnag("Avon Warwickshire", "MC", startyr = 2015, type = "River")}
+#' get_rnag("Avon Warwickshire", "MC", startyr = 2015, type = "River")
 #' 
 get_rnag <- function(ea_name = NULL, column = NULL, startyr = NULL, 
     endyr = NULL, type = NULL) {
