+++ man/search_names.Rd
@@ -23,9 +23,12 @@ Columns are defined in \code{\link{ea_wbids}}.
 Searches the listing of EA monitoring sites to find rows
 that contain the string provided. Can search by WBID (\code{WBID}), name 
 (\code{name}), Management Catchment (\code{MC}), Operational Catchment 
-(\code{OC}) or River Basin District (\code{RBD}). The search is done 
-on a local copy of the waterbody listing contained in the 
-\code{\link{ea_wbids}} object rather than connecting to the 
+(\code{OC}) or River Basin District (\code{RBD}). There is a hierarchical 
+relationship between these levels as shown at \url{https://environment.data.
+gov.uk/catchment-planning/help#help-catchment-hierarchy}.
+
+The search is done on a local copy of the waterbody listing contained in 
+the \code{\link{ea_wbids}} object rather than connecting to the 
 EA site.
 }
 \examples{
