+++ man/get_measures.Rd
@@ -35,9 +35,9 @@ the function will return an empty dataframe with a message.
 }
 \examples{
 # get the measures put in place for the Thames RBD
-\dontrun{get_measures("Thames", "RBD")}
+get_measures("Thames", "RBD")
 
 # get the measures put in place for the Loddon Operational Catchment
-\dontrun{get_measures("Loddon", "OC")}
+get_measures("Loddon", "OC")
 
 }
