+++ man/get_pa.Rd
@@ -33,10 +33,10 @@ Catchment (\code{OC}) or River Basin District (\code{RBD}).
 }
 \examples{
 # get protected areas associated with waterbody GB112071065700
-\dontrun{get_pa("GB112071065700", "WBID")}
+get_pa("GB112071065700", "WBID")
 
 # get the protected areas associated with the Humber RBD
-\dontrun{get_pa("Humber", "RBD")}
+get_pa("Humber", "RBD")
 
 # get the protected areas associated with the Avon Warwickshire
 # Management Catchment
