### 2019-05-07 4a4d5e4
updating ref table, plus correction in status message
