+++ vignettes/cde-output-reference.Rmd
@@ -31,49 +31,49 @@ operational_catchment | Operational Catchments are the next level down from Mana
 waterbody_id | Water Bodies are the smallest unit in the catchment hierarchy, representing all or part of a river system, lake, estuary etc. Each has a unique ID (WBID). | All functions
 status | WFD status classification (High, Good, Moderate, Poor or Bad). For objectives this represents the target (aimed for) status class. | All get_ functions except get_pa
 classification_level | The level within the range of classification elements that the status class refers to (NEEDS LINK) | objectives, rnag & status
-classification_level |  | -
-cycle |  | -
-water_body |  | -
-water_body_type |  | -
-year |  | -
+cycle | Which cycle of assessment the data come from. The first cycle. The first cycle ran from 2009 to 2015 and cycle 2 to 2015-2021. There were changes in some environmental standards between cycles and for theoverlap period, `cde` only returns cycle 2 assessments.  | -
+water_body | Name of the waterbody. | rnag etc.
+water_body_type | Type of waterbody. Values are: River, Lake, CoastalWater, TransitionalWater, GroundWaterBody | -
+year | Year to which the data refer to. | -
 classification_item |  | -
-easting |  | -
+easting | British National Grid Reference easting, i.e. x coordinate of location. | -
 hydromorphological_designation |  | -
-id |  | -
-investigation_outcome |  | -
-ngr |  | -
-northing |  | -
-activity_certainty |  | -
-business_sector |  | -
-category |  | -
-category_certainty |  | -
-certainty |  | -
-classification_element |  | -
-classification_id |  | -
-confidence |  | -
-estimated_start_date |  | -
-funding_stream |  | -
-lead_organisation |  | -
-linked_rnags |  | -
-measure_category_1 |  | -
-measure_category_2 |  | -
-measure_category_3 |  | -
-measure_reference_code |  | -
-measure_type |  | -
-national_swmi_header |  | -
-objectivetype |  | -
-pressure_tier_1 |  | -
-pressure_tier_2 |  | -
-pressure_tier_3 |  | -
-protected_area_code |  | -
-protected_area_label |  | -
-protected_area_type |  | -
-reason_type |  | -
-reasons_for_alternative_objectives |  | -
-rnag_url |  | -
-sector_of_lead_organisation |  | -
-see_also |  | -
-swmi |  | -
-swmi_certainty |  | -
-title |  | -
+id | Unique numerical code of RNAG (may be linked to measures specified).  | get_rnag
+investigation_outcome | Outcome of the investigation into the RNAG | -
+ngr | Ordnance Survey Landranger grid reference of the location. | -
+northing | British National Grid Reference northing, i.e. y coordinate of location.  | -
+activity | Classification of the activity linked to the RNAG identified, for example "Forestry", "Septic Tanks". | get_rnag
+activity_certainty | Assessment of the certainty with which the RNAG identified is linked to the activity specifed.  | -
+business_sector | Categorisation of the business sector linked to the activity causing the RNAG, for example "Transport", "Retail sector", "Forestry". | -
+category | Broad categorisation of the RNAG, for example "Industry", "Agriculture and rural land management" or "Water Industry". | -
+category_certainty | Assessment of the certainty with which the RNAG identified is linked to the category specifed. | get_rnag
+certainty | Assessment of the certainty of the status class assignment. | get_status
+classification_element | Specific quality element or metric linked to the RNAG. | get_rnag
+classification_id |  | ### NEEDED?? ###
+confidence | Confidence in the status class assignment. | get_status
+estimated_start_date | Estimated date at which action specified will being  | get_measures
+funding_stream | Source of funding for measure specified. | get_measures
+lead_organisation | Organisation leading the work on the measure specified. | get_measures
[…]
