### 2019-05-07 f952745
removing uri columns from objectives output
