### 2019-05-07 5698b20
removing classification_id column from get_status
