### 2019-05-07 dd0b543
adding on message for when fread takes time processing zipfile
