+++ vignettes/cde-output-reference.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "cde-output-reference"
+title: "cde: Reference list of all output columns"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{cde-output-reference}
@@ -15,44 +15,39 @@ knitr::opts_chunk$set(
 ```
 
 ```{r setup}
-library(cde)
+#library(cde)
 ```
+The table below gives information on all the current columns of data retrieved by the different `get_...` functions of `cde`; names, what they mean and which functions return them. The EA do modify the format of their data from time to time, so this will be updated as necessary.
 
-## Reference list of all output columns
-
-The table below gives information on all the current columns of data retrieved by the different functions of `cde`; names, what they mean and which functions return them. The EA do modify the format of their data from time to time, so this will be updated as necessary.
-
-
-Column name | Description | Output by which functions
+Column name | Description | Output from which functions
 --- | --- | ---
-river_basin_district | Name of the River Basin District. This is the top level in the [catchment hierarchy](https://environment.data.gov.uk/catchment-planning/help#help-catchment-hierarchy). | All functions
-management_catchment | Management Catchments are the next level down the hierarchy, see link above. | All functions
-operational_catchment | Operational Catchments are the next level down from Management Catchments. | All functions
-waterbody_id | Water Bodies are the smallest unit in the catchment hierarchy, representing all or part of a river system, lake, estuary etc. Each has a unique ID (WBID). | All functions
-status | WFD status classification (High, Good, Moderate, Poor or Bad). For objectives this represents the target (aimed for) status class. | All get_ functions except get_pa
-classification_level | The level within the range of classification elements that the status class refers to (NEEDS LINK) | objectives, rnag & status
-cycle | Which cycle of assessment the data come from. The first cycle. The first cycle ran from 2009 to 2015 and cycle 2 to 2015-2021. There were changes in some environmental standards between cycles and for theoverlap period, `cde` only returns cycle 2 assessments.  | -
-water_body | Name of the waterbody. | rnag etc.
-water_body_type | Type of waterbody. Values are: River, Lake, CoastalWater, TransitionalWater, GroundWaterBody | -
-year | Year to which the data refer to. | -
-classification_item |  | -
-easting | British National Grid Reference easting, i.e. x coordinate of location. | -
-hydromorphological_designation |  | -
-id | Unique numerical code of RNAG (may be linked to measures specified).  | get_rnag
-investigation_outcome | Outcome of the investigation into the RNAG | -
-ngr | Ordnance Survey Landranger grid reference of the location. | -
-northing | British National Grid Reference northing, i.e. y coordinate of location.  | -
+river_basin_district | Name of the River Basin District. This is the top level in the [catchment hierarchy](https://environment.data.gov.uk/catchment-planning/help#help-catchment-hierarchy). | All get_... functions
+management_catchment | Name of the Management Catchment, the next level down the hierarchy, see link above. | All get_... functions
+operational_catchment | name of the Operational Catchment, the next level down from Management Catchments. | All get_... functions
+waterbody_id | Waterbodies are the smallest unit in the catchment hierarchy, representing all or part of a river system, lake, estuary etc. Each has a unique ID (WBID). | All get_... functions
+status | WFD status classification (High, Good, Moderate, Poor or Bad). For objectives this represents the target (aimed for) status class. | All get_... functions except get_pa
+classification_item | Broad category of classification level (ecological, chemical or quantitative). | get_objectives, get_rnag & get_status
[…]
