### 2019-05-07 65719f0
reinstating plot and print tests
