### 2019-05-09 a76720e
updating vignette and column reference list
