+++ vignettes/cde-output-reference.Rmd
@@ -28,7 +28,7 @@ waterbody_id | Waterbodies are the smallest unit in the catchment hierarchy, rep
 status | WFD status classification (High, Good, Moderate, Poor or Bad). For objectives this represents the target (aimed for) status class. | All get_... functions except get_pa
 classification_item | Broad category of classification level (ecological, chemical or quantitative). | get_objectives, get_rnag & get_status
 classification_level | The level within the range of classification elements that the status class refers to (see details [here](https://environment.data.gov.uk/catchment-planning/help#help-surface-waters-classification-hierarchy)). | get_objectives, get_rnag & get_status
-cycle | Which cycle of assessment the data come from. The first cycle. The first cycle ran from 2009 to 2015 and cycle 2 to 2015-2021. There were changes in some environmental standards between cycles and for the overlap period, `cde` only returns cycle 2 assessments. | get_objectives, get_rnag & get_status
+cycle | Which cycle of assessment the data come from. The first cycle. The first cycle ran from 2009 to 2015 and cycle 2 to 2015-2021. There were changes in some environmental standards between cycles and for a two year period (2013-2014) where both assessment methods were employed, `cde` only returns cycle 2 assessments. | get_objectives, get_rnag & get_status
 water_body | Name of the waterbody. | get_objectives, get_rnag & get_status
 water_body_type | Type of waterbody. Values are: River, Lake, CoastalWater, TransitionalWater, GroundWaterBody | get_objectives, get_rnag & get_status
 year | Year to which the data refer to. | get_objectives, get_rnag & get_status
@@ -37,17 +37,17 @@ hydromorphological_designation | Hydromorphological designation of the waterbody
 ngr | Ordnance Survey Landranger grid reference of the location. | get_objectives & get_status
 northing | British National Grid Reference northing, i.e. y coordinate of location.  | get_objectives & get_status
 activity | Classification of the activity linked to the RNAG identified, for example "Forestry", "Septic Tanks". | get_rnag
-activity_certainty | Assessment of the certainty with which the RNAG identified is linked to the activity specifed.  | get_rnag
+activity_certainty | Categorical assessment of the certainty with which the RNAG identified is linked to the activity specifed.  | get_rnag
 business_sector | Categorisation of the business sector linked to the activity causing the RNAG, for example "Transport", "Retail sector", "Forestry". | get_rnag
 category | Broad categorisation of the RNAG, for example "Industry", "Agriculture and rural land management" or "Water Industry". | get_rnag
-category_certainty | Assessment of the certainty with which the RNAG identified is linked to the category specifed. | get_rnag
-certainty | Assessment of the certainty of the status class assignment. | get_status
+category_certainty | Categorical assessment of the certainty with which the RNAG identified is linked to the category specifed. | get_rnag
+certainty | Categorical assessment of the certainty of the status class assignment. | get_status
 classification_element | Specific quality element or metric linked to the RNAG. | get_rnag
 confidence | Confidence in the status class assignment (0-1 scale). | get_status
-estimated_start_date | Estimated date at which action specified will being  | get_measures
+estimated_start_date | Estimated date at which measure specified will begin.  | get_measures
 funding_stream | Source of funding for measure specified. | get_measures
 id | Unique numerical code of RNAG (may be linked to measures specified).  | get_rnag
-investigation_outcome | Outcome of the investigation into the RNAG | get_rnag
+investigation_outcome | Outcome of the investigation into the RNAG. | get_rnag
 lead_organisation | Organisation leading the work on the measure specified. | get_measures
[…]
