+++ vignettes/cde.Rmd
@@ -30,7 +30,7 @@ The CDE data are made available under the [Open Government Licence v3.0](https:/
 
 ## Installation
 
-You can install cde from github with:
+You can install `cde` from github with:
 
 ```{r gh-installation, eval = FALSE}
 # if you have not done so already
@@ -50,10 +50,11 @@ library(cde)
 lark_wb<-search_names(string="Lark", column="name")
 ```
 
-The dataframe returned contains details of all the waterbodies containing the string "Lark". The details include waterbody id codes (WBID), type of waterbody, Operational and Management Catchment names and River Basin District.
+The dataframe returned contains details of all the waterbodies containing the string "Lark" in their name. The details include waterbody id codes (WBID), type of waterbody, Operational and Management Catchment names and River Basin District.
 
 ```{r show lark_wb content,eval=TRUE}
-head(lark_wb)
+# show the top 6 rows of the 'name' column
+head(lark_wb$name)
 ```
 
 To search for Operational Catchments containing the same string we would use the following code.
@@ -92,17 +93,18 @@ head(lark_wb)
 lark_hawstead<-get_status(ea_name="GB105033042920", column="WBID")
 
 # the dataframe returned contains all of the data for this site in all 
-# years (we did not specify year/year range) (truncate to fit screen)
-str(lark_hawstead, vec.len=1, nchar.max=20)
+# years (we did not specify year/year range).
+lark_hawstead
 
 # just a quick look at the status data
-table(lark_hawstead$Status)
+table(lark_hawstead$status)
 ```
 An example of a higher level download, specifying a year range and type (in this case Rivers).
 
 ```{r lark river,eval=TRUE}
 # download status data for a given year range and type of waterbody
 lark_OC_rivers<-get_status(ea_name="Lark", column="OC", startyr=2013, endyr=2015, type="River")
+lark_OC_rivers
 ```
 To get information about status classification in relation to a specific level in the classification, we can specify `level` as well (see table above for options).
 
@@ -111,22 +113,27 @@ To get information about status classification in relation to a specific level i
 lark_OC_rivers_chem<-get_status(ea_name="Lark", column="OC", type="River", level="Chemical")
 ```
 
-The function returns a dataframe containing the status (and other details) for the specified combination of column, value, level and dates. Note that during 2013 and 2014 waterbodies were classified under both Cycle 1 and Cycle 2 methodologies. The status information extracted for these years is just for the Cycle 2 classification, to avoid double counting. There was also a change in some of the environmental standards applied to chemical aspects of status assessment between cycles, so there may be some noticeable changes in status between these years. See [here](https://environment.data.gov.uk/catchment-planning/help#help-surface-water-chemical-classification) for more details.
+The function returns an object of class `cde_df` (basically a dataframe with custom print and plot methods) containing the status (and other details) for the specified combination of column, value, level and dates. Note that during 2013 and 2014 waterbodies were classified under both Cycle 1 and Cycle 2 methodologies. The status information extracted for these years is just for the Cycle 2 classification, to avoid double counting. There was also a change in some of the environmental standards applied to chemical aspects of status assessment between cycles, so there may be some noticeable changes in status between these years. See [here](https://environment.data.gov.uk/catchment-planning/help#help-surface-water-chemical-classification) for more details.
+
+For details of the meaning of the the different columns, see the reference list.
 
 ## Plotting quality status classification data
 
[…]
