+++ man/ea_wbids.Rd
@@ -15,7 +15,10 @@
   \item{MC_num}{Index number of the Management Catchment}
   \item{RBD}{River Basin District name}
   \item{RBD_num}{Index number of the River Basin District}
-}}
+}
+For details of the hierarchy of the different catchment types, see
+\url{https://environment.data.gov.uk/catchment-planning/help#help-
+catchment-hierarchy}}
 \source{
 \url{https://environment.data.gov.uk/catchment-planning/}
 }
