### 2019-05-09 1d26956
big docs and pkgdown site update
