+++ man/cde.Rd
@@ -19,6 +19,5 @@ data downloaded (form of plot depends on data type).
 The website that is accessed is: 
 \url{https://environment.data.gov.uk/catchment-planning/}.
 The data are made available under the Open Government Licence v3.0 
-\url{https://www.nationalarchives.gov.uk/doc/open-government-licence
-/version/3/}
+\url{https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}
 }
