+++ man/ea_wbids.Rd
@@ -17,8 +17,7 @@
   \item{RBD_num}{Index number of the River Basin District}
 }
 For details of the hierarchy of the different catchment types, see
-\url{https://environment.data.gov.uk/catchment-planning/help#help-
-catchment-hierarchy}}
+\url{https://environment.data.gov.uk/catchment-planning/help#help-catchment-hierarchy}}
 \source{
 \url{https://environment.data.gov.uk/catchment-planning/}
 }
