+++ R/get_rnag.r
@@ -44,6 +44,9 @@
 #' @return An object of class \code{cde_df} containing the details of the 
 #' Reasons for Not Achieving Good Status for the specified combination 
 #' of criteria.
+#' For details of the meaning of the the different columns returned, 
+#' see \url{https://robbriers.github.io/cde/articles/cde-output-reference.html}.
+
 #'
 #' @export get_rnag
 #'
