+++ R/get_measures.r
@@ -22,6 +22,9 @@
 #'
 #' @return An object of class \code{cde_df} containing the details 
 #' of the measures put in place to try and improve water quality.
+#' For details of the meaning of the the different columns returned, 
+#' see \url{https://robbriers.github.io/cde/articles/cde-output-reference.html).
+
 #'
 #' @export get_measures
 #'
