+++ R/cde.r
@@ -1,7 +1,7 @@
 #' cde: Download Water Framework Directive (WFD) data from the 
 #' Environment Agency Catchment Data Explorer (CDE) website.
 #'
-#' The cde package facilitates searching and download of the WFD-related
+#' The `cde` package facilitates searching and download of the WFD-related
 #' data for all waterbodies within the Environment Agency area (i.e. England).
 #' The types of data that can be downloaded are: WFD status classification 
 #' data, Reasons for Not Achieving Good (RNAG) status, objectives set for 
@@ -12,7 +12,8 @@
 #' The website that is accessed is: 
 #' \url{https://environment.data.gov.uk/catchment-planning/}.
 
-#' The data are made available under the Open Government Licence v3.0 
+#' The data accessed by and included within the package are made available 
+#' under the Open Government Licence v3.0 
 #' \url{https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}
 #'
 #' @docType package
