+++ man/get_rnag.Rd
@@ -40,6 +40,8 @@ waterbody type. Possible values are \code{River}, \code{Lake},
 An object of class \code{cde_df} containing the details of the 
 Reasons for Not Achieving Good Status for the specified combination 
 of criteria.
+For details of the meaning of the the different columns returned, 
+see \url{https://robbriers.github.io/cde/articles/cde-output-reference.html}.
 }
 \description{
 Retrieves details of Reasons for Not Achieving Good (RNAG)
