+++ vignettes/test.Rmd
@@ -1,65 +1,0 @@
----
-title: "cde: testing"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{cde-testing}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-## Reference list of columns output
-
-The table below gives information on all the current columns of data retrieved by the different `get_...` functions of `cde`; names, what they mean and which functions return them. The EA do modify the format of their data from time to time, so this will be updated as necessary.
-
-Column name | Description | Output from which functions
---- | --- | ---
-river_basin_district | Name of the River Basin District. This is the top level in the [catchment hierarchy](https://environment.data.gov.uk/catchment-planning/help#help-catchment-hierarchy). | All get_... functions
-management_catchment | Name of the Management Catchment, the next level down the hierarchy, see link above. | All get_... functions
-operational_catchment | name of the Operational Catchment, the next level down from Management Catchments. | All get_... functions
-waterbody_id | Waterbodies are the smallest unit in the catchment hierarchy, representing all or part of a river system, lake, estuary etc. Each has a unique ID (WBID). | All get_... functions
-status | WFD status classification (High, Good, Moderate, Poor or Bad). For objectives this represents the target (aimed for) status class. | All get_... functions except get_pa
-classification_item | Broad category of classification level (ecological, chemical or quantitative). | get_objectives, get_rnag & get_status
-classification_level | The level within the range of classification elements that the status class refers to (see details [here](https://environment.data.gov.uk/catchment-planning/help#help-surface-waters-classification-hierarchy)). | get_objectives, get_rnag & get_status
-cycle | Which cycle of assessment the data come from. The first cycle. The first cycle ran from 2009 to 2015 and cycle 2 to 2015-2021. There were changes in some environmental standards between cycles and for a two year period (2013-2014) where both assessment methods were employed, `cde` only returns cycle 2 assessments. | get_objectives, get_rnag & get_status
-water_body | Name of the waterbody. | get_objectives, get_rnag & get_status
-water_body_type | Type of waterbody. Values are: River, Lake, CoastalWater, TransitionalWater, GroundWaterBody | get_objectives, get_rnag & get_status
-year | Year to which the data refer to. | get_objectives, get_rnag & get_status
-easting | British National Grid Reference easting, i.e. x coordinate of location. | get_objectives & get_status
-hydromorphological_designation | Hydromorphological designation of the waterbody (e.g. whether the body is artificial or heavily modified). | get_objectives & get_status
-ngr | Ordnance Survey Landranger grid reference of the location. | get_objectives & get_status
-northing | British National Grid Reference northing, i.e. y coordinate of location.  | get_objectives & get_status
-activity | Classification of the activity linked to the RNAG identified, for example "Forestry", "Septic Tanks". | get_rnag
-activity_certainty | Categorical assessment of the certainty with which the RNAG identified is linked to the activity specifed.  | get_rnag
-business_sector | Categorisation of the business sector linked to the RNAG, for example "Transport", "Retail sector", "Forestry" | get_rnag
-category | Broad categorisation of the RNAG, for example "Industry", "Agriculture and rural land management". | get_rnag
-
-certainty | Categorical assessment of the certainty of the status class assignment. | get_status
-classification_element | Specific quality element or metric linked to the RNAG. | get_rnag
-confidence | Confidence in the status class assignment (0-1 scale). | get_status
-estimated_start_date | Estimated date at which measure specified will begin.  | get_measures
-funding_stream | Source of funding for measure specified. | get_measures
[…]
