+++ man/get_pa.Rd
@@ -22,6 +22,8 @@ for specific values.}
 \value{
 An object of class \code{cde_df} containing the details of the 
 Protected Areas associated with the waterbodies.
+For details of the meaning of the the different columns returned, 
+see \url{https://robbriers.github.io/cde/articles/cde-output-reference.html}.
 }
 \description{
 Retrieves details of Protected Areas associated with 
