### 2019-05-10 01ef9d3
adding test to improve coverage of print
