+++ R/get_rnag.r
@@ -101,7 +101,7 @@ get_rnag <- function(ea_name = NULL, column = NULL, startyr = NULL,
       }
     }
     # subset data
-    rnag_data<-subset_data(rnag_data, column, NULL, startyr, endyr, type)
+    rnag_data<-subset_data(rnag_data, column, NULL, startyr, endyr, type, data_type=NULL)
     if (nrow(rnag_data)==0){
       message("No RNAG data - empty dataframe returned")
     }
