### 2019-05-10 ee0d793
reworking obj check in subset function
