### 2019-05-10 6373fc1
fixes for date range subsetting and extra tests to get this sucker over 90%
