### 2019-05-10 d707b6f
updating JOSS paper to reflect changes
