### 2019-05-10 19b8738
final update to docs (for now!)
