+++ R/search_names.r
@@ -24,10 +24,10 @@
 #'
 #' @examples
 #' # search for River Basin Districts containing "Tweed"
-#' \dontrun{search_names("Tweed", "RBD")}
+#' \dontrun{search_names(string="Tweed", column="RBD")}
 #' 
 #' # search for Management Catchments containing "Kennet"
-#' \dontrun{search_names("Kennet", "MC")}
+#' \dontrun{search_names(string="Kennet", column="MC")}
 #' 
 search_names <- function(string = NULL, column = NULL) {
   search_choices <- c("name","WBID", "MC", "OC", "RBD")
