+++ R/get_status.r
@@ -67,18 +67,18 @@
 #'
 #' @examples
 #' # get Overall Water Body status classification for waterbody GB520804714300
-#' get_status("GB520804714300", "WBID")
+#' get_status(ea_name="GB520804714300", column="WBID")
 #' 
 #' # get status class based on Priority substances for waterbody GB520804714300
-#' get_status("GB520804714300", "WBID", level = "Priority substances")
+#' get_status(ea_name="GB520804714300", column="WBID", level="Priority substances")
 #' 
 #' # get the Overall Water Body status of Lakes in the Humber RBD, between
 #' # 2012 and 2014
-#' get_status("Humber", "RBD", startyr = 2012, endyr = 2014, type = "Lake")
+#' get_status(ea_name="Humber", column="RBD", startyr=2012, endyr=2014, type="Lake")
 #' 
 #' # get the Overall Water Body status for Rivers in the Avon Warwickshire
 #' # Operational Catchment in 2011
-#' get_status("Avon Warwickshire", "MC", startyr = 2011, type = "River")
+#' get_status(ea_name="Avon Warwickshire", column="MC", startyr=2011, type="River")
 #' 
 get_status <- function(ea_name = NULL, column = NULL, 
     level = "Overall Water Body", startyr = NULL, endyr = NULL, type = NULL) {
