+++ man/get_measures.Rd
@@ -37,9 +37,9 @@ the function will return an empty dataframe with a message.
 }
 \examples{
 # get the measures put in place for the Thames RBD
-get_measures("Thames", "RBD")
+get_measures(ea_name="Thames", column="RBD")
 
 # get the measures put in place for the Loddon Operational Catchment
-get_measures("Loddon", "OC")
+get_measures(ea_name="Loddon", column="OC")
 
 }
