### 2019-05-10 d1641b4
typo in objectives docs
