### 2019-05-11 74dd250
more doc updates - issue with rendering of \tabular{} content in Rd (status and objectives)
