+++ R/get_objectives.r
@@ -28,12 +28,17 @@
 #' within the classification used, see \url{https://environment.data.gov.uk/catchment-planning/help#help-classification-hierarchy}.
 #' 
 #' Defaults to 'Overall Water Body'. Possible values for the different levels 
-#' retrived by the function are shown below.
+#' retrieved by the function are shown below.
+#' \tabular{ccc}{
+#' Level 1 \tab Level 2 \tab Level 3 \cr
+#' This is the first col \tab This is the second column \tab This is the third col \cr
+#' }
+#' 
 #' \tabular{ccc}{
 #'  \strong{Level 1} \tab \strong{Level 2} \tab \strong{Level 4}\cr
-#'  Ecological \tab Biological quality elements \tab Overall Water Body\cr
-#'Chemical \tab Chemical Status element \tab -\cr
-#'  Quantitative \tab Hydromorphological Supporting Elements \tab -\cr
+#' Ecological \tab Biological quality elements \tab Overall Water Body \cr
+#' Chemical \tab Chemical Status element \tab -\cr
+#' Quantitative \tab Hydromorphological Supporting Elements \tab -\cr
 #'  - \tab Other Substances \tab -\cr
 #'  - \tab Physico-chemical quality elements \tab -\cr
 #'  - \tab Priority hazardous substances \tab -\cr
@@ -68,7 +73,7 @@
 #' get_objectives(ea_name="Humber", column="RBD", year=2021, type="Lake")
 #' 
 #' # get the objectives set for Rivers in the Avon Warwickshire
-#' # Operational Catchment in relation to Chemical status
+#' # Management Catchment in relation to Chemical status
 #' get_objectives(ea_name="Avon Warwickshire", column="MC", level="Chemical", type="River")
 #' 
 get_objectives <- function(ea_name = NULL, column = NULL, 
