+++ man/search_names.Rd
@@ -31,10 +31,10 @@ the \code{\link{ea_wbids}} object rather than connecting to the
 EA site.
 }
 \examples{
-# search for River Basin Districts containing "Tweed"
-\dontrun{search_names(string="Tweed", column="RBD")}
+# search for sites containing "Tadnoll" in the name
+search_names(string="Tadnoll", column="name")
 
-# search for Management Catchments containing "Kennet"
-\dontrun{search_names(string="Kennet", column="MC")}
+# search for Operational Catchments containing "Cornwall"
+search_names(string="Cornwall", column="OC")
 
 }
