+++ man/get_objectives.Rd
@@ -26,10 +26,6 @@ within the classification used, see \url{https://environment.data.gov.uk/catchme
 
 Defaults to 'Overall Water Body'. Possible values for the different levels 
 retrieved by the function are shown below.
-\tabular{ccc}{
-Level 1 \tab Level 2 \tab Level 3 \cr
-This is the first col \tab This is the second column \tab This is the third col \cr
-}
 
 \tabular{ccc}{
  \strong{Level 1} \tab \strong{Level 2} \tab \strong{Level 4}\cr
