### 2019-05-13 e2bedeb
fix for tiny console width
