### 2019-05-13 9fb0be3
deleting old docs from previous functions.
