### 2019-05-13 #284 @robbriers
Ok, took a bit longer than planned, but here we go...
First @boshek review:
_`get_measures` function: There is an issue with the measures data on the CDE website since the last quarterly update - this is being actively worked on by the EA and should be fixed soon. This has been documented in the vignette. but at the moment all calls to this will return an empty dataframe._
Review Comments
I think this is a great package that leverage data access provided by the UK government. It is certainly within rOpenSci' scope for package. The documentation is very good and easy to follow which effectively lower the barrier to entry for a package user. I have some specific comments which are outlined below.

Documentation
On the blog down page, you are linked to the "Getting Started" section even though it is "Get Started". This is a potential source of confusion.
_This has been updated on the pkgdown pages to point to "Get Started"._
Vignette
I think you could revise this line

and use of this package implies acceptance of these licence conditions

to:

and use of the data accessed by this package implies acceptance of these licence conditions
_Revised and expanded to cover data contained within package as well._
Consider using remotes for install_github()
_Revised to use remotes_
Consider using full argument names in the getting started materials.
_All examples now use full argument names._
Code
Consider prefixing every function with cde_*
_Kept as they are as I think there is limited scope for confusion with other packages._
Consider returning columns as all lower case with underscores instead of decimals
_All column names now lower case with underscores._
This file: https://github.com/robbriers/cde/blob/master/R/sysdata.rda. Internal data in the package should be included in the data folder.
_Thanks for your help with this: internal data now included in the data folder and documented in terms of structure and content. Raw csv also available on Github repo in data-raw folder._
No LICENSE file. The open data license would license the data used but not the code in the package
Here is a good description of how to do this: http://r-pkgs.had.co.nz/data.html. In addition, that data should be license within the repo (if it ships with package).
Here is an example of that: https://github.com/ropensci/tidyhydat/tree/master/data-raw/HYDAT_internal_data
_LICENSE file added with appropriate details; documentation also updated to highlight licencing of data in several places._
This chunk of code: https://github.com/robbriers/cde/blob/d0e109a2794a2653b15410926a52d2a1f15469bf/R/utils.r#L37-L51
And others like it could probably benefit from a) some helper functions and b) use of switch.
_`switch` is my new best friend! The code highlighted (and several other places) have been broken up to use helper functions and reduce repetition._
This worries me: https://github.com/robbriers/cde/blob/d0e109a2794a2653b15410926a52d2a1f15469bf/R/utils.r#L94. Could you instead catch the specific error using tryCatch()?
tryCatch() now implemented to deal with the warning generated, see https://github.com/robbriers/cde/blob/9fb0be3ac62ce7aa5f9836c7c7c797632bd1e2eb/R/utils.r#L47-L50.
A helper function called something like cde_data.table or something like that would be good as you are replicating lots of code here. (https://github.com/robbriers/cde/blob/d0e109a2794a2653b15410926a52d2a1f15469bf/R/utils.r#L94-L114)
_Code has been broken up into a number of smaller helper functions to reduce replication, see https://github.com/robbriers/cde/blob/9fb0be3ac62ce7aa5f9836c7c7c797632bd1e2eb/R/utils.r#L59-L139_
I think you should consider using data.table's progress bar. When downloads are happening, it is ambiguous to the user if the process is working.
_Profiling the code has highlighted that it is the download and unzip parts of the process where most of the latency is. `curl` has been replaced by `download.file` as the messages given are more informative, and a message also added to highlight continued activity when unzipping is occurring. I have also set `showProgress=TRUE` for the data.table part, but this was rarely needed on any of the machines that I tested it on._
Consider developing print methods (and a class) for cde objects that only print out 10 rows (or just import tibble and use those print methods)
I think this could also benefit from plot and maybe even summary methods.
_Custom `print` and `plot` methods associated with a new `cde_df` class have been implemented for all output from `get_`... functions. `search_names` has been retained as a plain df as truncation of values to fit the console width may result in confusion if using the data in the `get_`... functions._
Consider letting your examples run so that they are tested with CI
_All examples now run._
