### 2019-05-15 f2d7915
adding rOpensci footer
