+++ README.Rmd
@@ -39,3 +39,5 @@ remotes::install_github("robbriers/cde")
 ## Basic usage
 
 Details of how to use the package can be found at [https://robbriers.github.io/cde](https://robbriers.github.io/cde).
+
+[![ropensci_footer](https://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
