+++ index.Rmd
@@ -39,3 +39,5 @@ remotes::install_github("robbriers/cde")
 ## Basic usage
 
 See the 'Get started' vignette or 'Reference' sections above for details of the different functions.
+
+[![ropensci_footer](https://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
