### 2019-05-15 50634f8
updating links to ropensci, still need to rebuild site
