+++ index.Rmd
@@ -33,7 +33,7 @@ You can install cde from github with:
 
 ```{r gh-installation, eval = FALSE}
 # install.packages("remotes")
-remotes::install_github("robbriers/cde")
+remotes::install_github("ropensci/cde")
 ```
 
 ## Basic usage
