+++ man/get_measures.Rd
@@ -23,7 +23,7 @@ specific values.}
 An object of class \code{cde_df} containing the details 
 of the measures put in place to try and improve water quality.
 For details of the meaning of the the different columns returned, 
-see \url{https://robbriers.github.io/cde/articles/cde-output-reference.html}.
+see \url{https://ropensci.github.io/cde/articles/cde-output-reference.html}.
 }
 \description{
 Retrieves details of the measures put in place or proposed 
