+++ man/get_rnag.Rd
@@ -41,7 +41,7 @@ An object of class \code{cde_df} containing the details of the
 Reasons for Not Achieving Good Status for the specified combination 
 of criteria.
 For details of the meaning of the the different columns returned, 
-see \url{https://robbriers.github.io/cde/articles/cde-output-reference.html}.
+see \url{https://ropensci.github.io/cde/articles/cde-output-reference.html}.
 }
 \description{
 Retrieves details of Reasons for Not Achieving Good (RNAG)
