+++ .github/CONTRIBUTING.md
@@ -34,7 +34,7 @@ GitHub username, and links to relevant issue(s)/PR(s).
 
 ### Code of Conduct
 
-Please note that the {{{ package }}} project is released with a
+Please note that the `cde` project is released with a
 [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this
 project you agree to abide by its terms.
 
