### 2019-05-15 #284 @stefaniebutland
@robbriers Here are the blog post guidelines https://github.com/ropensci/roweb2#contributing-a-blog-post

Happy to answer any questions.
